pyproject.toml
[project]
name = "lab1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "pyyaml>=6.0.3",
]
